add CORS setting
